Cpp odities
