+++ src/managers/animation/AttachPoint.cpp
@@ -20,3 +20,33 @@ namespace {
     log::info("Complete");
   }
 }
+
+namespace Gts {
+  Actor* GetActorPtr(Actor* actor) {
+    return actor;
+  }
+
+  Actor* GetActorPtr(Actor& actor) {
+    return &actor;
+  }
+
+  Actor* GetActorPtr(ActorHandle& actor) {
+    if (!actor) {
+      return nullptr;
+    }
+    return actor.get().get();
+  }
+  Actor* GetActorPtr(const ActorHandle& actor) {
+    if (!actor) {
+      return nullptr;
+    }
+    return actor.get().get();
+  }
+  Actor* GetActorPtr(FormID formId) {
+    Actor* actor = TESForm::LookupByID<Actor>(formId);
+    if (!actor) {
+      return nullptr;
+    }
+    return actor;
+  }
+}
