+++ src/managers/animation/AttachPoint.hpp
@@ -5,33 +5,15 @@ using namespace RE;
 
 namespace Gts {
 
-  Actor* GetActorPtr(Actor* actor) {
-    return actor;
-  }
+  Actor* GetActorPtr(Actor* actor);
 
-  Actor* GetActorPtr(Actor& actor) {
-    return &actor;
-  }
+  Actor* GetActorPtr(Actor& actor);
 
-  Actor* GetActorPtr(ActorHandle& actor) {
-    if (!actor) {
-      return nullptr;
-    }
-    return actor.get().get();
-  }
-  Actor* GetActorPtr(const ActorHandle& actor) {
-    if (!actor) {
-      return nullptr;
-    }
-    return actor.get().get();
-  }
-  Actor* GetActorPtr(FormID formId) {
-    Actor* actor = TESForm::LookupByID<Actor>(formId);
-    if (!actor) {
-      return nullptr;
-    }
-    return actor;
-  }
+  Actor* GetActorPtr(ActorHandle& actor);
+
+  Actor* GetActorPtr(const ActorHandle& actor);
+
+  Actor* GetActorPtr(FormID formId);
 
   template<typename T, typename U>
   bool AttachTo(T& anyGiant, U& anyTiny, NiPoint3 point) {
